requirements.txt
asgiref==3.4.1
click==8.0.3
colorama==0.4.4
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-mpd2==3.0.4
Werkzeug==2.0.2
